Implement checkpoint/restore commands
This enables the checkpoint/restore commands. Right now it can be used
like this:

 # crictl checkpoint CTR-ID
 # crictl restore CTR-ID

To checkpoint and restore a container. It also offers the possibility to
restore a container in another pod:

 # crictl checkpoint CTR-ID
 # crictl restore -p POD-ID CTR-ID # POD-ID being the destination pod

Both presented use case so far are not very useful but necessary
implementation steps towards the actual use case:

  # crictl checkpoint --export=/tmp/cp.tar CTR-ID
  # reboot
  # crictl runp pod.json # to create new pod
  # crictl restore --import=/tmp/cp.tar --pod=podID

This enables rebooting of a system without losing the state of a
container.

In addition to checkpointing containers it is also possible (with CRI-O)
to checkpoint Pods using a Pod ID instead of a container ID:

  # crictl checkpoint --export=/tmp/cp.tar POD-ID

To restore this Pod following command can be used:

  # crictl restore --import=/tmp/cp.tar

CRI-O will now create a new Pod and restore all containers into this new
Pod.

Signed-off-by: Adrian Reber <areber@redhat.com>